refactor(desktop): move chat model selection helpers
